Refactor Question Logic
