+++ src/components/PlacePicker.tsx
@@ -37,7 +37,7 @@ import {
     determineName,
     geocode,
     type OpenStreetMap,
-} from "@/maps/api/index";
+} from "@/maps/api";
 
 import { Button } from "./ui/button";
 import { Separator } from "./ui/separator";
